CSW PROFILES: Profiles service & bean (cont).
