+++ geoportal-connectors/geoportal-harvester-csw/src/main/java/com/esri/geoportal/harvester/csw/CswConnector.java
@@ -18,8 +18,6 @@ package com.esri.geoportal.harvester.csw;
 import static com.esri.geoportal.commons.constants.CredentialsConstants.*;
 import static com.esri.geoportal.harvester.csw.CswConstants.*;
 import com.esri.geoportal.commons.csw.client.IProfile;
-import com.esri.geoportal.commons.csw.client.IProfiles;
-import com.esri.geoportal.commons.csw.client.impl.ProfilesProvider;
 import com.esri.geoportal.commons.csw.client.impl.ProfilesService;
 import com.esri.geoportal.harvester.api.defs.EntityDefinition;
 import com.esri.geoportal.harvester.api.defs.UITemplate;
@@ -39,12 +37,14 @@ import java.util.ResourceBundle;
  */
 public class CswConnector implements InputConnector<InputBroker> {
   public static final String TYPE = "CSW";
-  private final IProfiles profiles;
   private final ProfilesService profilesService;
 
   public CswConnector(ProfilesService profilesService, String cswProfilesFolder) {
     this.profilesService = profilesService;
-    this.profiles = new ProfilesProvider(cswProfilesFolder).newProfiles();
+  }
+  
+  public ProfilesService getProfilesService() {
+    return profilesService;
   }
  
   @Override
@@ -68,10 +68,10 @@ public class CswConnector implements InputConnector<InputBroker> {
         return true;
       }
     });
-    Choice<String>[] choices = profiles.listAll().stream().map(p->new Choice<String>(p.getId(),p.getName())).toArray(Choice[]::new);
+    Choice<String>[] choices = profilesService.newProfiles().listAll().stream().map(p->new Choice<String>(p.getId(),p.getName())).toArray(Choice[]::new);
     arguments.add(new UITemplate.ChoiceArgument(P_PROFILE_ID, bundle.getString("csw.profile"), Arrays.asList(choices)){
       public String getDefault() {
-        IProfile defaultProfile = profiles.getDefaultProfile();
+        IProfile defaultProfile = profilesService.newProfiles().getDefaultProfile();
         return defaultProfile!=null? defaultProfile.getId(): null;
       }
     });
@@ -81,11 +81,11 @@ public class CswConnector implements InputConnector<InputBroker> {
 
   @Override
   public void validateDefinition(EntityDefinition definition) throws InvalidDefinitionException {
-    new CswBrokerDefinitionAdaptor(profiles, definition);
+    new CswBrokerDefinitionAdaptor(profilesService.newProfiles(), definition);
   }
 
   @Override
   public InputBroker createBroker(EntityDefinition definition) throws InvalidDefinitionException {
-    return new CswBroker(this, new CswBrokerDefinitionAdaptor(profiles, definition));
+    return new CswBroker(this, new CswBrokerDefinitionAdaptor(profilesService.newProfiles(), definition));
   }
 }
