+++ geoportal-connectors/geoportal-harvester-migration/src/main/java/com/esri/geoportal/harvester/migration/MigrationConnector.java
@@ -39,6 +39,10 @@ public class MigrationConnector implements InputConnector<InputBroker>{
     this.profilesService = profilesService;
   }
 
+  public ProfilesService getProfilesService() {
+    return profilesService;
+  }
+
   @Override
   public void validateDefinition(EntityDefinition definition) throws InvalidDefinitionException {
     new MigrationBrokerDefinitionAdaptor(definition);
